Submitting PFFNY SPL Token name, symbol, and icon
PocketFuel or PFFNY (Pronounced 'Piffany'), is the governance token that incentivizes community growth and engagement on the PocketFuel.io DFN of Decentralized Freelance Network. It powers the NFT royalty payments system and drives our funding and recruitment crowdfunding. Inspired by the word Epiphany, or realization, the token symbol is an acronym for PocketFuel.io Freelance Network Yields.